Used Uint32 for SDL_Delay call

+++ GAME_NAME/Motor2D/j1App.cpp
@@ -437,7 +437,7 @@ void j1App::FramerateLogic() {
 	if (mustCapFPS) {
 		int delayTime = (1000 / fpsCap) - lastFrameMs;
 		if (delayTime > 0) {
-			SDL_Delay(delayTime);
+			SDL_Delay((Uint32)delayTime);
 			LOG("We waited for %u and got back in %f", delayTime, delayTimer.ReadMs());
 		}
 	}
